==> your-project/Code/3_Create_Inventory_Master_File_from_Transactions.ipynb <==
##Libraries
#Dataframe & Arrays
import pandas as pd
import numpy as np



df = pd.read_csv('../data/dataframe_awp_staging.csv')

df = df.sort_values(by=['StockCode','UnitPrice'], ascending=False)

df_pricing_file = df.groupby(['B2B','StockCode','UnitPrice']).agg({'Rev': ['sum','count'],'Quantity':['sum']}).reset_index()
df_pricing_file = df_pricing_file.sort_values(by=['StockCode','UnitPrice'], ascending=False)
df_pricing_file.head()

df_pricing_file.B2B = df_pricing_file.B2B.astype(int)

df_pricing_file.columns = ['B2B','StockCode','UnitPrice','Rev','Count','Qty']
df_pricing_file.head()

df_awp = df_pricing_file.groupby(['B2B','StockCode']).agg({'Rev': ['sum'],'Qty':['sum']})
df_awp.columns = ['Rev','Qty']

df_awp.head()

def AWP(row):
    return (row['Rev'] / row['Qty'])

df_awp['AWP'] = (df_awp['Rev'] / df_awp['Qty'])
df_awp.head()

df_awp.reset_index(inplace=True)
df_awp.columns  = ['B2B','StockCode','Rev','Qty','AWP']
df_awp.head()

## trying to identify average  Qty purchased by StockCode and customer type

df_qty_mean = df.groupby(['B2B','StockCode'])['Quantity'].mean().astype(int).to_frame()
df_qty_mean.reset_index(inplace=True)
df_qty_mean.columns  = ['B2B','StockCode','Quantity']

df_qty_mean.head()

def qtymeanlookup(row):
    """
    Input: StockCode taken from row within dataframe
    Output: The corresponding mean quantity sold for that item
    
    The purpose of this funtion is to vlookup value of item from a reference dataframe (df_qty_mean)
    """
    try:

        return int(df_qty_mean.loc[(df_qty_mean['StockCode'] == row.StockCode) & (df_qty_mean['B2B'] == row.B2B)]["Quantity"])

    except:

        pass

df_awp['MeanQty'] = df_awp.apply(qtymeanlookup, axis=1)

df_awp.head()

df_awp.to_csv('../data/df_awp.csv', index=False)

#####################################################
#####################################################
# Create - a different pricing file without B2B Flag
#####################################################
#####################################################

df_pricing_file_noB2B = df.groupby(['StockCode','UnitPrice']).agg({'Rev': ['sum','count'],'Quantity':['sum']}).reset_index()
df_pricing_file_noB2B = df_pricing_file_noB2B.sort_values(by=['StockCode','UnitPrice'], ascending=False)
df_pricing_file_noB2B.head()

df_pricing_file_noB2B.columns = ['StockCode','UnitPrice','Rev','Count','Qty']
df_pricing_file_noB2B.head()

df_awp_noB2B = df_pricing_file_noB2B.groupby(['StockCode']).agg({'Rev': ['sum'],'Qty':['sum']})
df_awp_noB2B.columns = ['Rev','Qty']

df_awp_noB2B.head()

df_awp_noB2B['AWP'] = (df_awp_noB2B['Rev'] / df_awp_noB2B['Qty'])
df_awp_noB2B.head()

df_awp_noB2B.reset_index(inplace=True)
df_awp_noB2B.columns  = ['StockCode','Rev','Qty','AWP']
df_awp_noB2B.head()

df_qty_mean_noB2B = df.groupby(['StockCode'])['Quantity'].mean().astype(int).to_frame()
df_qty_mean_noB2B.reset_index(inplace=True)
df_qty_mean_noB2B.columns  = ['StockCode','Quantity']

df_qty_mean_noB2B.head()

def qtymeanlookup_noB2B(row):
    """
    Input: StockCode taken from row within dataframe
    Output: The corresponding mean quantity sold for that item
    
    The purpose of this funtion is to vlookup value of item from a reference dataframe (df_qty_mean)
    """
    try:

        return int(df_qty_mean_noB2B.loc[(df_qty_mean_noB2B['StockCode'] == row.StockCode) ]["Quantity"])

    except:

        pass

df_awp_noB2B['MeanQty'] = df_awp_noB2B.apply(qtymeanlookup_noB2B, axis=1)

df_awp_noB2B.head()

df_awp_noB2B.to_csv('../data/df_awp_noB2B.csv', index=False)
