==> your-project/Code/4_Recommender_System.ipynb <==
# Recommender System | Revenue Potential 

This notebook contains all the code to product product recommendations, estimate the total revenue over 9 months and then annualise it.

### Loading Ncessary Libraries and Defining Functions

Getting the notebook ready for this part of the project.

##Libraries
#Dataframe & Arrays
import pandas as pd
import numpy as np

#SCiPy for Cluster Distance Analysis
from scipy.spatial.distance import pdist, squareform

#Required for Function within Function
from functools import partial

def customertypeinrecsys(row):
    """
    Input: CustomerID
    Output: Flag to identify if B2B customer or not (1 or 0)
    
    The purpose of this funtion is to vlookup if the customer ID belongs to a business or regular customer
    """
    try:
        return int(df_custtype.loc[df_custtype['CustomerID'] == row]["B2B"])
    except:
        return 0

#Working version
def pricelookup(row, col):
    """
    Input: StockCode taken from row within dataframe
    Output: The corresponding value of that item
    
    The purpose of this funtion is to vlookup value of item from a reference dataframe (df_price_lookup)
    """
    if row[col] is not None:
        try:
            awp = float(df_awp.loc[(df_awp['StockCode'] == row[col]) & (df_awp['B2B'] == row['B2B'])]["AWP"])
            mean = float(df_awp.loc[(df_awp['StockCode'] == row[col]) & (df_awp['B2B'] == row['B2B'])]["MeanQty"])
            return awp*mean
        except:
            awp2 = float(df_awp_noB2B.loc[(df_awp_noB2B['StockCode'] == row[col]) ]["AWP"])
            mean2 = float(df_awp_noB2B.loc[(df_awp_noB2B['StockCode'] == row[col])]["MeanQty"])
            return float(awp2*mean2)
    else:
        return 0

### Loading Necessary Dataframes from CSV Files

The files were created in a different notebook.

#Load main dataframe require for this project
#df = pd.read_csv('../data/dataframe_recom_system.csv')

#Load main dataframe required for this project Updates to row above as uses different source file
df = pd.read_csv('../data/dataframe_recom_system_NR.csv')

#Load the dataframe that contains the Average Weighted Prices and Quantity Purchased by Customer Type (B2B or B2C)
df_awp = pd.read_csv('../data/df_awp.csv')

#Load the dataframe that contains the Average Weighted Prices and Quantity Purchased overall 
#Required if can't be found in above file
df_awp_noB2B = pd.read_csv('../data/df_awp_noB2B.csv')

df_awp.dtypes

#Ensure the B2B Flag is an Integer (1 or 0)
df_awp['B2B'] = df_awp['B2B'].astype(int)

#Ensure there are no Nans in CustomerID fields and values are integers
df['CustomerID'] = df['CustomerID'].astype(int)
df['CustomerID'].isna().sum()

df.shape

### Recommender System

First to determine the total number of items (StockCodes) purchased by each customer (CustomerID)

df_cust_prod_grouped = pd.DataFrame(df.groupby(['CustomerID', 'StockCode'])['Quantity'].agg('sum'))
df_cust_prod_grouped.head()

Next to create a pivot table of customer against stock code to show quantity of each item purchased.

#Applying pivot table to df created above (df_cust_prod_grouped).
df_cust_matrix = df_cust_prod_grouped.pivot_table('Quantity', 'StockCode', 'CustomerID', aggfunc='sum', fill_value = 0)
df_cust_matrix.head()

Now to create a matrix that provides info on the similarity of customers.

# I need to transpose the matrix, otherwise I get the distance for products, not customers.
# The pdist function returns a 1-Dimensional array.
# Then applying squareform returns a vector-form distance vector
# Finally convert it into a DataFrame
#The "euclidean" metric will be used for this example.

df_distance_matrix = pd.DataFrame(squareform(pdist(df_cust_matrix.T, metric='euclidean')), \
                                  index=df_cust_matrix.columns, columns=df_cust_matrix.columns)
df_distance_matrix.head()

# Given the values are so large and hard to determine their meaning they will be inverted and normalised. 
# ie. the closer to 1, the more similar they are

df_distance_matrix_norm = pd.DataFrame(1/(1 + df_distance_matrix))
df_distance_matrix_norm.head()

It's time to check the output of the above by generating a list of the top 5 most similar customers for a specific CustomerID.

df_top5_recs_spec_customer = df_distance_matrix_norm[12350].sort_values(ascending = False).head(6)
df_top5_recs_spec_customer

For one of the similar customers identified above, it's time to see what they purchased.

df_sample_similar_cust_recomms = df_cust_prod_grouped.loc[(df_top5_recs_spec_customer.index[1:],)]
df_sample_similar_cust_recomms.head()

The dataframe above will now be used to create and aggregate StockCode, and Quantity. Thwy will then be ordered in descending order of quantity.

This will provide the total number of each product purchased by the 5 most similar customers to the customer in descending order of quantity purcahsed. (The assumption is that if the Quantity purcahsed is higher then it's more like to be appealign to a similar customer).

df_agg_similar_prod_ranked = df_sample_similar_cust_recomms.groupby('StockCode')[['Quantity']].sum()\
                .sort_values(by = 'Quantity', ascending = False)
df_agg_similar_prod_ranked.head()

Next, it's time to filter the list for products that the chosen customer has not yet purchased and then recommend the top 5 products with the highest quantities remaining.

df_product_recomms = pd.concat([df_agg_similar_prod_ranked, df_cust_matrix[12350]], axis=1, sort=False)
df_product_recomms.rename(columns = {12350:'test_customer'}, inplace = True)
df_product_recomms

df_top5_recomm = df_product_recomms.query('Quantity > 0 and test_customer == 0').head(5)
df_top5_recomm

The recommendations made for a single customer now need to be applied across all customers.

dict_recommendations = {}
unique_ID = df_distance_matrix_norm.columns.unique()

for customer in unique_ID:
    head = df_distance_matrix_norm[customer].sort_values(ascending = False).head(6)
    df_sample_similar_cust_recomms = df_cust_prod_grouped.loc[(head.index[1:],)]
    df_agg_similar_prod_ranked = df_sample_similar_cust_recomms.groupby('StockCode')[['Quantity']].sum()\
                .sort_values(by = 'Quantity', ascending = False)
    df_product_recomms = pd.concat([df_agg_similar_prod_ranked, df_cust_matrix[customer]], axis=1, sort=False)
    df_product_recomms.rename(columns = {customer:'customer'}, inplace = True)
    dict_recommendations[customer] = list(df_product_recomms.query('Quantity > 0 and customer == 0').head(5).index)
    

The above results now need to be converted into a dataframe (from dictionary).

df_recommendations = pd.DataFrame.from_dict(dict_recommendations, orient='index', 
                                columns=['rec1', 'rec2', 'rec3', 'rec4', 'rec5'])
df_recommendations.head()

## Putting it all together...
This section will run all the code for the different distance metrics availble and compare them all.

metrics = [ 'cityblock', 'correlation', 'cosine', 'dice', 
           'euclidean', 'hamming', 'jaccard']


for i in metrics:
    df_distance_matrix = pd.DataFrame(squareform(pdist(df_cust_matrix.T, metric=i)), index=df_cust_matrix.columns, columns=df_cust_matrix.columns)
    df_distance_matrix_norm = pd.DataFrame(1/(1 + df_distance_matrix))
    dict_recommendations = {}
    unique_ID = df_distance_matrix_norm.columns.unique()

    for customer in unique_ID:
        head = df_distance_matrix_norm[customer].sort_values(ascending = False).head(6)
        df_sample_similar_cust_recomms = df_cust_prod_grouped.loc[(head.index[1:],)]
        df_agg_similar_prod_ranked = df_sample_similar_cust_recomms.groupby('StockCode')[['Quantity']].sum()\
                    .sort_values(by = 'Quantity', ascending = False)
        df_product_recomms = pd.concat([df_agg_similar_prod_ranked, df_cust_matrix[customer]], axis=1, sort=False)
        df_product_recomms.rename(columns = {customer:'customer'}, inplace = True)
        dict_recommendations[customer] = list(df_product_recomms.query('Quantity > 0 and customer == 0').head(5).index)

    df_recommendations = pd.DataFrame.from_dict(dict_recommendations, orient='index', 
                                    columns=['rec1', 'rec2', 'rec3', 'rec4', 'rec5'])    

    #drop the index
    df_recommendations.reset_index(drop=False, inplace=True)

    #rename the index
    df_recommendations.columns = ['CustomerID','rec1','rec2','rec3','rec4','rec5']

    #add B2B columns
    df_recommendations['B2B'] = df_recommendations['CustomerID'].apply(customertypeinrecsys)

    #lookup prices based on the stock code and B2B flag
    for column in ['rec1','rec2','rec3','rec4','rec5']:
        look = partial(pricelookup, col=column)
        df_recommendations[f"{column}value"] = df_recommendations.apply(look, axis=1)
    
    
    
    df_recommendations['TotalPossCustRev'] = df_recommendations.iloc[:,-5:].sum(axis=1)

    Total = df_recommendations['TotalPossCustRev'].sum()
    Total_Annual = ((Total/11)*12)

    print(f"The total for {i} metric was : {Total} for 11 months or {Total_Annual} annualised")
    #print("The total for {} metric was : {0:.2f} for 11 months or {0:.2f} annualised".format(i, Total,Total_Annual ))

print("All done")

The total for cityblock metric was : 370758.48566564394 for 9 months or 404463.80254433886 annualised
The total for correlation metric was : 315393.3666921541 for 9 months or 344065.49093689537 annualised
The total for cosine metric was : 315750.65981281555 for 9 months or 344455.26525034424 annualised
The total for dice metric was : 349029.88932659273 for 9 months or 380759.87926537386 annualised
The total for euclidean metric was : 310853.27841711615 for 9 months or 339112.6673641267 annualised
The total for hamming metric was : 423977.1892219803 for 9 months or 462520.57006034214 annualised
The total for jaccard metric was : 335560.0019931393 for 9 months or 366065.4567197883 annualised
All done
