==> your-project/Code/5_Calculate_Returns_Percentage.ipynb <==

The percentage of merchandise returned over a particular period can be calculated by simply dividing the number of items returned by the number that has been sold

#Libraries
#Dataframe and arrays
import pandas as pd
import numpy as np

#Determine how many of the units that were sold during the time period were later returned, 
#and divide that number by the number of units sold. Multiply your answer by 100 to calculate the percentage of units returned.

df = pd.read_csv('../data/data_for_returns_analysis.csv')

df['Rev'] = df['Quantity']*df['UnitPrice']

df_returns = df.loc[df['Quantity'] < 0]

df['Return'] = df.apply(lambda row: 1 if row.Quantity < 0 else 0, axis=1)

df.head()

df_returns.head()

df['Sold'] = df.apply(lambda row: row.Quantity if row.Quantity > 0 else 0, axis=1)

df['Returned'] = df.apply(lambda row: row.Quantity if row.Quantity < 0 else 0, axis=1)

df.head()


df_return_rate = df.groupby('StockCode').agg({'Sold': ['sum'],'Returned': ['sum']})
df_return_rate.reset_index(inplace=True)


df_return_rate.columns = ['StockCode','Sold','Returned']


df_return_rate.head()

#df_return_rate['return_percent_unit'] = df.apply(lambda row: (row.Returned / row.Sold)*100, axis=1)

df_return_rate['return_percent_unit'] = df.apply(lambda row: ((row.Returned / row.Sold)*100) if row.Sold != 0 else 0, axis=1)


df_return_rate.head()

df_return_rate['return_percent_unit'] = ( df_return_rate['Returned'] / df_return_rate['Sold'] ) * -100

df_return_rate.sort_values(by='return_percent_unit', ascending=False).head(20)

#df.loc[df.StockCode=='47351B'] #BLUE VOILE LAMPSHADE	
#df.loc[df.StockCode=='90176A'] #CLASSIC DIAMANTE NECKLACE JET		
#df.loc[df.StockCode=='84247K'] #FOLK ART GREETING CARD,PACK/12		
#df.loc[df.StockCode=='22105'] #MIRROR MOSAIC GOBLET CANDLE HOLDER
#df.loc[df.StockCode=='22102'] #MIRROR MOSAIC T-LIGHT HOLDER
#df.loc[df.StockCode=='22103'] #MIRROR MOSAIC T-LIGHT HOLDER ROUND
#df.loc[df.StockCode=='22920'] #HERB MARKER BASIL
