==> your-project/Code/6_Categorise_Individual_Business_Customer.ipynb <==
# Criteria to Determine if Customer is An Individual or Business/ Organisation

The recommender system is going to work best when seperating individuals from businesses given that businesses are likely to be less persuaded with retargeting and other digital ad campaigns.

#Libraries
#Dataframe and arrays
import pandas as pd
import numpy as np

import matplotlib as plt


#make sure the clean file is passed in. 
df = pd.read_excel('../data/online_fashion.xlsx')

df['Rev'] = df['Quantity']* df['UnitPrice']

df.head()

df.dropna(subset=['CustomerID'],inplace=True)
df_new.dropna(subset=['CustomerID'],inplace=True)

df_new.sort_values(by='UnitPrice', ascending=False)

#Any customer buying over 2500 worth of goods should be categorised as business

df_cust_value = df.groupby('CustomerID').agg({'Rev': ['sum']}).sort_values(by=[('Rev','sum')], ascending=False) #nested column names
df_cust_value.reset_index(inplace=True)
df_cust_value.columns = ['CustomerID','TotalRev']
df_cust_value.head()

df_cust_value.loc[(df_cust_value['TotalRev'] < 2500) & (df_cust_value['TotalRev'] > 0)]

#df.loc[(df['column_name'] >= A) & (df['column_name'] <= B)]

###Identify customers spending over 2500 - DONE

df_cust_items = df.groupby('CustomerID').agg({'Quantity': ['sum']}).sort_values(by=[('Quantity','sum')], ascending=False) #nested column names
df_cust_items.reset_index(inplace=True)
df_cust_items.columns = ['CustomerID','TotalNoItem']
df_cust_items.head()

df_cust_items.loc[(df_cust_items['TotalNoItem'] < 60) & (df_cust_items['TotalNoItem'] > 0)].head()

df_cust_items.loc[(df_cust_items['TotalNoItem'] > 60)]

df_new.sort_values(by='UnitPrice', ascending=False)
