==> your-project/Code/7_Clustering_Exercise.ipynb <==
#Import Required Libraries
import pandas as pd
import numpy as np

#Import geocoder
from geopy import Nominatim
import geopy.distance

#Import Machine Learning Libraries
from sklearn.decomposition import PCA
from sklearn.cluster import KMeans

df = pd.read_csv('../data/dataframe_recom_system_NR.csv')

df.head(1)

df.CustomerID = df.CustomerID.astype(int)

df.columns

df.drop(['Unnamed: 0'], axis=1, inplace=True)

df.describe()

df.head(1)

# Adding the Countries to Get Distance from UK

#Finding Customers with more than 1 Country listed
df_customer_stg = df.groupby('CustomerID')['Country'].unique()
custs2drop = df_customer_stg.loc[df_customer_stg.apply(lambda x:len(x)>1)]

df_custs2drop = pd.DataFrame(custs2drop)

df_custs2drop = df_custs2drop.reset_index()

custs2drop_list = df_custs2drop['CustomerID'].tolist()

df.shape

df[~df['CustomerID'].isin(custs2drop_list)].shape

df_clean = df[~df['CustomerID'].isin(custs2drop_list)]

cuntries = []
for i in df['Country'].unique():
    cuntries.append(i)

df_count = pd.DataFrame(cuntries)

df_count.head()

df_count.columns = ['Country']

df_count.head()

df_count['Latitude'] = df_count['Country'].apply(geolocator.geocode).apply(lambda x: x.latitude)
df_count['Longitude'] = df_count['Country'].apply(geolocator.geocode).apply(lambda x: x.longitude)

df_count.isna().sum()


df_count[df_count['Country'] == 'UNITED KINGDOM']


UK = (54.702354, -3.276575)


#Iterating between customers' countries to find the distance from UK
cust_country = range(df_count.shape[0])
distances = []
for s in cust_country:
    location = (df_count.iloc[s]['Latitude'],df_count.iloc[s]['Longitude'])
    distances.append(round(geopy.distance.vincenty(UK, location).km,2))

#Adding distances as feature
df_count['UKDist'] = distances   

df_count.sort_values(by='UKDist')


df_count.head(1)

# Obtaining to Features to Cluster By

df_summary = df.groupby(['CustomerID','Country']).agg({'Rev': ['sum','mean','min','max'],'InvoiceNo':'count','Quantity':'sum'})

df_summary.head()

#I've not reset the index so that CustomerID and Country dont get mixed up in clustering
df_summary = df_summary.reset_index()
df_summary.head()

df_summary.columns = ['CustomerID','Country','Rev_Total','TransAvgVal','Trans_Min','Trans_Max','Trans_Num','TotQty']

df_summary.head()

df_final_stg = df_summary.merge(df_count, on='Country', how='left')

df_final_stg.head(10)

df_final = df_final_stg[['CustomerID','Rev_Total','TransAvgVal','Trans_Min','Trans_Max','Trans_Num','TotQty','UKDist']]

df_final.head(10)

df_final.set_index('CustomerID', inplace=True)

df_final.head(1)

# Given the values are so large and hard to determine their meaning they will be inverted and normalised. 
# ie. the closer to 1, the more similar they are

df_final = pd.DataFrame(1/(1 + df_final))
df_final.head()


## Model Build: Part 1 - K-means
The dataframe is now ready to build the model, containing all the required features. Initially, I will start with 6 clusters - a low, middle and high category for each B2B and B2C customers.

kmeans = KMeans(n_clusters=6)
customer_clusters = kmeans.fit(df_final)
customer_clusters.cluster_centers_

## Model Build: Part 2 - PCA
With 7 features - it's worth comparing k-means with PCA.

# Principal Component 
pca = PCA(n_components=6)

principalComponents = pca.fit_transform(df_final)
principalDf = pd.DataFrame(data = principalComponents
             ,columns = ['pc1', 'pc2', 'pc3', 'pc4','pc5', 'pc6'])
principalDf.head()

kmeans = KMeans(n_clusters=6)
october_clusters_PCA = kmeans.fit(principalDf)
october_clusters_PCA.cluster_centers_

## Model Addition to DataFrame

df_final['labels'] = customer_clusters.fit_predict(df_final)
df_final.head()

df_final['labelsPCA'] = customer_clusters.fit_predict(principalDf)
df_final.head()

df_final[['labels','labelsPCA']] = df_final[['labels','labelsPCA']].replace([0, 1, 2, 3, 4, 5], [1, 2, 3, 4, 5, 6])

df_final.labels.value_counts()

df_final.labelsPCA.value_counts()

#Let's check if there is any relationship between the labels created by the two models:
df_final.groupby(["labels","labelsPCA"]).count()
